refactor(ui): share liquid glass filter pipeline
Extract the duplicated measurement, displacement-map, and SVG filter logic used by liquid glass surfaces. Keep the public barrel limited to supported component and physics exports.
