+++ src/components/shared/liquid-glass/index.ts
@@ -1,19 +1,10 @@
 export { LiquidGlass } from './LiquidGlass';
 export { LiquidGlassLayer } from './LiquidGlassLayer';
 export {
-  clearLensMapCache,
-  generateLensMap,
-  getLensMapDataUrl,
-  type LensMapResult,
-} from './generate-lens-map';
-export {
   DEFAULT_LIQUID_GLASS_PHYSICS,
   MARKETING_CTA_PHYSICS,
   MARKETING_HEADER_PHYSICS,
   PRICING_HEADER_PHYSICS,
   resolveLiquidGlassPhysics,
-  type LiquidGlassLens,
-  type LiquidGlassLayerProps,
   type LiquidGlassPhysics,
-  type LiquidGlassProps,
 } from './types';
